fix: 代码审查修复 — 5个CRITICAL + 8个HIGH + 8个MEDIUM问题
CRITICAL修复:
- master.py: 修复evidence_chains导致TypeError崩溃（免费用户推命全部失败）
- payments.py: 修复价格标签显示¥298/¥49与实际¥365/¥59不符
- middleware.ts: 修复Host header注入导致的开放重定向漏洞
- payments.py: /paypal/return端点增加用户认证防止未授权订单激活
- QRPaymentModal.tsx: 修复window.open误用paypalOrderId作为URL

HIGH修复:
- prompts.py: evidence_chains参数现在正确注入到LLM prompt中
- graph.py: 英文报告不再显示中文硬编码文本
- auth.py: X-Real-IP仅信任本地代理，防止限速绕过
- payments.py: 移除未使用常量(GRANT_ON_REGISTER, CNY_TO_USD_RATE)
- lunarCalendar.ts: 修复公元前年份数组索引为负数
- models.py: amount_usdt从Float改为Numeric(12,2)避免精度问题
- cron.yaml: secrets改为env变量而非内联shell注入
- auth.py: 替换print为logger

MEDIUM修复:
- QRPaymentModal.tsx: 支付验证错误不再静默吞掉
- QRPaymentModal.tsx: 轮询超时增加用户提示
- checkout/page.tsx: 增加防重复提交保护
- cron.py: 修复Bearer token解析使用split替代replace
- product_matcher.py: 修复None值导致TypeError
- product_matcher.py: print替换为logger

+++ backend/database/models.py
@@ -7,6 +7,7 @@ from __future__ import annotations
 import enum
 import uuid
 from datetime import datetime
+from decimal import Decimal
 from typing import Optional
 
 from sqlalchemy import (
@@ -524,7 +525,7 @@ class CryptoOrder(Base):
     user_id: Mapped[str] = mapped_column(
         ForeignKey("users.id", ondelete="CASCADE"), nullable=False, index=True
     )
-    amount_usdt: Mapped[float] = mapped_column(Float, nullable=False)
+    amount_usdt: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=False)
     network: Mapped[str] = mapped_column(String(20), nullable=False)  # "TRC20" | "ARBITRUM"
     stardust_granted: Mapped[int] = mapped_column(Integer, default=0)
     status: Mapped[str] = mapped_column(String(20), default="pending")  # pending|success|failed
